store page using the view item
